requirements.txt
# Core dependencies
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
itsdangerous==2.2.0
requests==2.32.3
urllib3==2.4.0

# Flask web framework
Flask==3.1.1
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3

# Data processing
numpy==2.2.6
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Computer Vision
opencv-python==4.11.0.86
pillow==11.2.1

# AI/ML
torch==2.7.0
torchvision==0.22.0
ultralytics==8.3.146
ultralytics-thop==2.0.14
thop==0.1.1.post2209072238

# Scientific computing
scipy==1.15.3
matplotlib==3.10.3
seaborn==0.13.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
pyparsing==3.2.3

# YOLOv5 dependencies
GitPython==3.1.44
gitdb==4.0.12
smmap==5.0.2
fsspec==2025.5.1
PyYAML==6.0.2
tqdm==4.67.1
packaging==25.0

# Hardware interface (Raspberry Pi)
RPi.GPIO==0.7.1
setuptools<81.0.0

# System utilities
psutil==7.0.0
py-cpuinfo==9.0.0
watchdog==6.0.0

# Environment management
python-dotenv==1.1.0

# Math utilities
mpmath==1.3.0
networkx==3.4.2
sympy==1.14.0
typing_extensions==4.13.2
six==1.17.0

# WSGI HTTP Server
gunicorn
